requirements.txt
Flask==3.0.0
face-recognition==1.3.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
gspread==5.12.0
google-auth-oauthlib==1.0.0
google-auth==2.23.4
Pillow==10.1.0
pandas==2.1.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==3.0.1
